Move autodispose android artifacts to autodispose2.* package
