+++ android/autodispose-android-archcomponents-test/gradle.properties
@@ -17,4 +17,4 @@
 POM_NAME=AutoDispose (Android Architecture Components Test Extensions)
 POM_ARTIFACT_ID=autodispose-android-archcomponents-test
 POM_PACKAGING=aar
-AUTOMATIC_MODULE_NAME=com.uber.autodispose.android.lifecycle.test
+AUTOMATIC_MODULE_NAME=autodispose2.android.lifecycle.test
