+++ android/autodispose-android-archcomponents-test/src/main/AndroidManifest.xml
@@ -14,4 +14,4 @@
   ~ limitations under the License.
   -->
 
-<manifest package="com.uber.autodispose.android.lifecycle.test"/>
+<manifest package="autodispose2.android.lifecycle.test"/>
