+++ android/autodispose-android-archcomponents/consumer-proguard-rules.txt
@@ -1,4 +1,4 @@
 -keep class * implements androidx.lifecycle.GeneratedAdapter {<init>(...);}
 
 # Top-level functions that can only be used by Kotlin.
--dontwarn com.uber.autodispose.android.lifecycle.KotlinExtensions
+-dontwarn autodispose2.android.lifecycle.KotlinExtensions
