+++ android/autodispose-android-archcomponents/gradle.properties
@@ -17,4 +17,4 @@
 POM_NAME=AutoDispose (Android Architecture Components Extensions)
 POM_ARTIFACT_ID=autodispose-android-archcomponents
 POM_PACKAGING=aar
-AUTOMATIC_MODULE_NAME=com.uber.autodispose.android.lifecycle
+AUTOMATIC_MODULE_NAME=autodispose2.android.lifecycle
