--- android/autodispose-android/src/androidTest/java/com/uber/autodispose/android/AutoDisposeTestActivity.java
@@ -1,35 +1,0 @@
-/*
- * Copyright (C) 2019. Uber Technologies
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.uber.autodispose.android;
-
-import android.app.Activity;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.FrameLayout;
-
-@SuppressWarnings("NullAway") // https://github.com/uber/NullAway/issues/19
-public final class AutoDisposeTestActivity extends Activity {
-  FrameLayout parent;
-  View child;
-
-  @Override
-  protected void onCreate(Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-    parent = new FrameLayout(this);
-    child = new View(this);
-    setContentView(parent);
-  }
-}
