+++ android/autodispose-android/src/main/AndroidManifest.xml
@@ -14,4 +14,4 @@
   ~ limitations under the License.
   -->
 
-<manifest package="com.uber.autodispose.android"/>
+<manifest package="autodispose2.android"/>
