+++ sample/src/main/java/com/uber/autodispose/recipes/AutoDisposeView.java
@@ -21,7 +21,7 @@ import android.util.AttributeSet;
 import android.view.View;
 import androidx.annotation.Nullable;
 import androidx.annotation.RequiresApi;
-import com.uber.autodispose.android.ViewScopeProvider;
+import autodispose2.android.ViewScopeProvider;
 import com.uber.autodispose.lifecycle.CorrespondingEventsFunction;
 import com.uber.autodispose.lifecycle.LifecycleEndedException;
 import com.uber.autodispose.lifecycle.LifecycleScopeProvider;
