+++ sample/src/main/kotlin/com/uber/autodispose/recipes/AutoDisposeViewKotlin.kt
@@ -20,7 +20,7 @@ import android.os.Build
 import android.util.AttributeSet
 import android.view.View
 import androidx.annotation.RequiresApi
-import com.uber.autodispose.android.ViewScopeProvider
+import autodispose2.android.ViewScopeProvider
 import com.uber.autodispose.lifecycle.CorrespondingEventsFunction
 import com.uber.autodispose.lifecycle.LifecycleEndedException
 import com.uber.autodispose.lifecycle.LifecycleScopeProvider
