+++ sample/src/main/kotlin/com/uber/autodispose/sample/KotlinActivity.kt
@@ -20,7 +20,7 @@ import android.util.Log
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.Lifecycle
 import autodispose2.autoDispose
-import com.uber.autodispose.android.lifecycle.AndroidLifecycleScopeProvider
+import autodispose2.android.lifecycle.AndroidLifecycleScopeProvider
 import com.uber.autodispose.autoDispose
 import com.uber.autodispose.recipes.subscribeBy
 import io.reactivex.rxjava3.core.Observable
