+++ sample/src/main/kotlin/com/uber/autodispose/sample/KotlinFragment.kt
@@ -23,7 +23,7 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Lifecycle
 import autodispose2.autoDispose
-import com.uber.autodispose.android.lifecycle.AndroidLifecycleScopeProvider
+import autodispose2.android.lifecycle.AndroidLifecycleScopeProvider
 import com.uber.autodispose.autoDispose
 import com.uber.autodispose.recipes.subscribeBy
 import io.reactivex.rxjava3.core.Observable
